can use an offset in annotations and spatial dists
This can be used to actually only compute data for the minimum
bounding box of the area that interests us (but still have positions
that are global to the origin of the atlas).

This speeds up substantially the computation of some datasets (orientations)
for some small-ish areas (hippocampus) since most of the values calculated
were zero anyways.

Change-Id: I8987d5b6274fe55c67b12e7d192362258acaee80