load recipe includes the layer percentage
which was being ignored

This shouldn't affect MOST of the cell mtype/etype/morphology assignment
because we are constructing the spatial distributions for those taking
into account the layers (effectively grouping by them), so the
re-normalization that happens at that point should remove this effect.

It will impact layer 2/3 which is treated separately in the BBP recipe
(with different proportions for mtypes and total number of cells) but
as a single region by the ABI.

On the other hand, not having it means you couldn't use the data from
the recipe to compare distributions of two or more layers together,
which we want to do in post-building analysis.

Change-Id: I9cf72b3a74f28c82c8096a786e13346e7f4f09b2