moved cell positioning code to voxcell
this makes voxcell more selfcontained, as different classes in it
were always talking about "positions" but we weren't providing
even a basic way to create these.

kept the module in brainbuilder as a thin wrapper, in the same
fashion as all the assign_* modules, so that future development
happens there.

once more advanced positioning develops into brainbuilder
(with cell avoidance for example), if we find that the algorithms
become generic (not workflow/bbp/region-specific), they can also
be moved into voxcell.

Change-Id: I39afd487101b42b5392edf1e5b034262e71a1707